fix: Student Step pipeline work with MFE
Make the Student Step pipeline MFE compatible.
This is a part of the Student Step pipeline fix for MFE.
MFE fires the tracking event with the `vertical_id` instead of `sequential_id`
(comparing to the legacy frontend).
Asked for the reason in the openedx MFE slack channel:
https://openedx.slack.com/archives/C01AJQU07R8/p1647419993954629

YT: https://youtrack.raccoongang.com/issue/RGOeX-991